% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_model_names.R
\name{find_model_names}
\alias{find_model_names}
\title{Find available names for a stock assessment model}
\usage{
find_model_names(data, model)
}
\arguments{
\item{data}{The StockSMART data.}

\item{model}{The assessment model to be checked. Current options are AMAK, ASAP, BAM, and SS}
}
\value{
A list of available model names
}
\description{
Find available names for a stock assessment model
}
\examples{
\dontrun{
find_model_names(
data = stocksmart_data,
model = "ASAP"
)
}
}
