% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_model_names.R
\name{standardize_model_names}
\alias{standardize_model_names}
\title{Add a new column with standardized model names}
\usage{
standardize_model_names(data, model, names)
}
\arguments{
\item{data}{The StockSMART data.}

\item{model}{The assessment model to be checked. Current options are AMAK, ASAP, BAM, and SS}

\item{names}{The available names for each model}
}
\value{
A new data frame of function \code{standardize_model_names}, containing a new column of Assessment.Model.Standardize.
}
\description{
Add a new column with standardized model names
}
\examples{
\dontrun{
standardize_model_names(
data = stocksmart_data,
model = "ASAP",
names = "asap"
)
}
}
